Testing: Refactor test_warnings to use its fixture for other tests
